pyproject.toml
[tool.poetry]
name = "dotenv_switch"
version = "0.0.1"
description = "Switch between sourcing .env files based on $APP_ENV environment variable"
readme = "README.md"
authors = ["Michael Berkowski <mjb@umn.edu>", "David Naughton <naughton@umn.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.3"
python_dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-forked = "*"
